configure eslint for ts + add rules
