+++ package.json
@@ -8,7 +8,9 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.2.0",
     "@testing-library/user-event": "^14.1.1",
-    "eslint-plugin-react-hooks": "^4.5.0",
+    "eslint-plugin-jest": "^27.6.0",
+    "eslint-plugin-prettier": "^5.0.1",
+    "eslint-plugin-typescript-sort-keys": "^3.1.0",
     "firebase": "^9.8.2",
     "immer": "^9.0.15",
     "prop-types": "^15.8.1",
@@ -63,12 +65,21 @@
     "@types/react-redux": "^7.1.24",
     "@types/styled-components": "^5.1.25",
     "@types/uuid": "^9.0.6",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.0.0",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
     "babel-plugin-styled-components": "^2.0.7",
-    "eslint": "^8.1.0",
-    "eslint-plugin-react": "^7.32.2",
+    "eslint": "^8.53.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "^17.1.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-cypress": "^2.15.1",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-unicorn": "^49.0.0",
     "gh-pages": "^4.0.0",
+    "prettier": "^3.0.3",
     "typescript": "^4.0.8",
     "typescript-eslint": "^0.0.1-alpha.0"
   },
