+++ src/App.test.js
@@ -5,11 +5,11 @@ import { store } from "./store";
 import App from "./App";
 
 test("renders learn react link", () => {
-  const { getByText } = render(
-    <Provider store={store}>
-      <App />
-    </Provider>
-  );
+    const { getByText } = render(
+        <Provider store={store}>
+            <App />
+        </Provider>
+    );
 
-  expect(getByText(/learn/i)).toBeInTheDocument();
+    expect(getByText(/learn/i)).toBeInTheDocument();
 });
