+++ src/components/EmptyList.tsx
@@ -1,0 +1,366 @@
+import styled from "styled-components";
+
+const EmptyListContainer = styled.div`
+  height: 100%;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 6.3rem;
+
+  @media (min-width: 768px) {
+    margin-top: 4rem;
+  }
+`;
+
+const SVGContainer = styled.svg`
+  align-self: center;
+  justify-content: center;
+  height: 200px;
+  width: 242px;
+`;
+
+const NothingHere = styled.h1`
+  margin-top: 4rem;
+  text-align: center;
+  font-weight: 700;
+  font-size: 20px;
+  line-height: 22px;
+  letter-spacing: -0.625px;
+
+  color: ${({ theme }) => theme.text};
+`;
+
+const Description = styled.p`
+  width: 242px;
+  font-weight: 500;
+  font-size: 12px;
+  line-height: 15px;
+  /* or 125% */
+
+  text-align: center;
+  letter-spacing: -0.25px;
+
+  color: ${({ theme }) => theme.greyText};
+
+  .newInvoice {
+    font-weight: 700;
+  }
+
+  .newInvoiceLargeOnly {
+    display: none;
+  }
+
+  @media (min-width: 768px) {
+    .newInvoiceLargeOnly {
+      display: inline;
+    }
+  }
+`;
+function EmptyList() {
+    return (
+        <EmptyListContainer>
+            <SVGContainer>
+                <svg width="242" height="200" xmlns="http://www.w3.org/2000/svg">
+                    <g fillRule="nonzero" fill="none">
+                        <path
+                            fill="#E6E7E8"
+                            d="M31.136 83.879l85.035-53.67 85.036 53.67v78.304H31.136z"
+                        />
+                        <path
+                            d="M201.207 162.577H31.136a.394.394 0 01-.394-.394V83.88a.378.378 0 01.181-.33L115.96 29.91a.402.402 0 01.417 0l85.067 53.677c.117.07.189.195.189.33v78.305a.394.394 0 01-.425.354zm-169.7-.787h169.345V84.1l-84.68-53.402-84.666 53.401v77.691z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M136.887 62.8c2.827-4.666 3.544-10.339 3.284-15.784a38.428 38.428 0 00-4.592-17.034 15.387 15.387 0 00-4.915-5.634c-2.552-1.574-4.293-.511-6.79-.936-2.11-.362-4.316-1.865-6.655-2.25a21.632 21.632 0 00-15.517 3.241C93.7 29.81 90.005 40.14 81.735 45.16c-8.105 4.894-18.478 3.423-27.568 6.13a30.73 30.73 0 00-10.523 5.555 18.346 18.346 0 00-5.151 6.145c-2.135 4.453-1.576 9.756 0 14.454 1.575 4.83 4.237 9.442 8.365 12.44 3.457 2.494 7.703 3.642 11.901 4.398 5.71 1.03 11.626 1.416 17.329.18a25.263 25.263 0 0011.397-5.539c1.402-1.25 2.174-2.958 3.481-4.201 1.804-1.708 4.616-2.754 6.727-4.06 3.3-2.054 6.608-4.241 10.373-5.217 9.562-2.486 23.055-3.163 28.82-12.644z"
+                            fill="#FFF"
+                        />
+                        <path
+                            d="M68.47 95.626a55.295 55.295 0 01-9.798-.96c-3.639-.653-8.294-1.754-12.05-4.453-3.758-2.699-6.68-7.082-8.492-12.59-1.291-3.933-2.363-9.732.056-14.72a18.408 18.408 0 015.245-6.295 31.037 31.037 0 0110.642-5.618 64.453 64.453 0 0111.208-1.92c5.876-.645 11.429-1.259 16.265-4.178 4.222-2.55 7.302-6.61 10.287-10.536 2.827-3.721 5.758-7.569 9.664-10.228a21.987 21.987 0 0115.753-3.329c1.254.265 2.471.683 3.624 1.244.98.458 2.008.809 3.064 1.046.798.112 1.607.13 2.41.055a6.748 6.748 0 014.505.937 15.498 15.498 0 015.041 5.751 38.523 38.523 0 014.64 17.2c.307 6.452-.788 11.802-3.34 15.98-4.726 7.727-14.398 9.702-22.952 11.449-2.159.44-4.19.85-6.089 1.345-3.395.874-6.467 2.793-9.452 4.65l-.788.512c-.787.464-1.575.897-2.402 1.322a18.647 18.647 0 00-4.261 2.698 13.442 13.442 0 00-1.575 1.944 13.175 13.175 0 01-1.907 2.266 25.476 25.476 0 01-11.61 5.68 35.702 35.702 0 01-7.687.748zm45.157-74.37c-4.17-.022-8.25 1.2-11.72 3.509-3.82 2.573-6.703 6.373-9.5 10.048-3.024 3.981-6.15 8.104-10.49 10.716-4.97 3.006-10.863 3.651-16.542 4.28-3.74.3-7.446.932-11.074 1.889a30.383 30.383 0 00-10.397 5.508 17.803 17.803 0 00-5.057 6.019c-2.284 4.768-1.268 10.347 0 14.163 1.765 5.42 4.608 9.654 8.223 12.258 3.616 2.605 8.2 3.698 11.76 4.344 6.632 1.188 12.24 1.25 17.163.18a24.743 24.743 0 0011.28-5.507 12.592 12.592 0 001.803-2.148 13.796 13.796 0 011.678-2.046 19.206 19.206 0 014.426-2.817c.843-.433 1.63-.842 2.363-1.29l.835-.52c3.017-1.88 6.136-3.831 9.641-4.72 1.922-.496 3.938-.913 6.12-1.354 8.42-1.707 17.95-3.65 22.488-11.11l.307.19-.307-.19c2.441-4.028 3.528-9.268 3.23-15.579a37.736 37.736 0 00-4.546-16.877 14.891 14.891 0 00-4.789-5.508 5.984 5.984 0 00-4.088-.787c-.858.073-1.722.05-2.575-.07a15.07 15.07 0 01-3.214-1.094 15.229 15.229 0 00-3.434-1.157 21.405 21.405 0 00-3.584-.33z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M169.023 55.09l-4.143 15.578 9.98 1.778 5.686-20.811s-8.869-6.782-10.24-6.83c-1.37-.047-1.283 10.284-1.283 10.284z"
+                            fill="#FFF"
+                        />
+                        <path
+                            d="M174.86 72.84h-.071l-9.964-1.755a.386.386 0 01-.307-.488l4.111-15.508c0-1.668 0-9.316 1.166-10.441a.678.678 0 01.52-.212c1.41.047 8.955 5.767 10.46 6.908a.401.401 0 01.142.417l-5.695 20.811a.386.386 0 01-.362.268zm-9.452-2.463l9.184 1.62 5.513-20.213c-4.26-3.241-9.034-6.562-9.79-6.585-.56.417-.906 5.508-.875 9.914a.4.4 0 010 .102l-4.032 15.162z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M158.24 44.9s1.41 3.147 3.442 3.659c2.032.511 6.451-1.669 6.963-1.385.512.283 12.17 9.001 13.303 9.26 1.135.26 3.222-2.486 4.112-5.751 0 0-11.342-7.727-12.083-8.545-.74-.818-1.622-5.712-2.725-7.506L158.24 44.9z"
+                            fill="#FFF"
+                        />
+                        <path
+                            d="M182.067 56.852a1.301 1.301 0 01-.245 0c-.708-.173-4.426-2.81-10.61-7.318-1.393-1.023-2.606-1.896-2.756-1.99-.15-.095-.945.283-1.575.503-1.647.543-3.9 1.29-5.27.929-2.19-.583-3.639-3.754-3.702-3.887a.377.377 0 01.11-.465l12.989-10.228a.41.41 0 01.315-.079c.11.023.206.088.268.181a21.34 21.34 0 011.575 4.383c.237 1.06.602 2.088 1.087 3.06.52.575 7.467 5.39 12.012 8.49.141.091.205.264.157.425-.811 2.864-2.82 5.996-4.355 5.996zM168.51 46.757a.654.654 0 01.323.07c.165.087.992.693 2.843 2.038 3.151 2.274 9.586 6.995 10.327 7.192.74.197 2.615-1.975 3.568-5.209-1.914-1.298-11.209-7.656-11.918-8.45-.401-.44-.74-1.574-1.268-3.368a27.7 27.7 0 00-1.284-3.8l-12.374 9.78c.386.74 1.576 2.777 3.049 3.147 1.15.3 3.363-.433 4.82-.912a7.584 7.584 0 011.914-.488z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M170.638 37.567a4.221 4.221 0 00.449 2.415c.638 1.141 3.465 2.66 3.465 2.66l1.97-3.934-3.86-2.786-2.024 1.645z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M174.552 43.035a.363.363 0 01-.18-.047c-.308-.158-2.955-1.613-3.624-2.817a4.614 4.614 0 01-.496-2.66.378.378 0 01.134-.243l2.024-1.668a.394.394 0 01.488 0l3.852 2.785c.154.11.207.316.126.488l-1.97 3.934a.394.394 0 01-.235.197.308.308 0 01-.119.031zm-3.536-5.256a3.805 3.805 0 00.417 2.014c.41.787 2.04 1.786 2.954 2.314l1.63-3.29-3.34-2.415-1.661 1.377z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M182.988 48.598s1.032-3.25-.788-3.478a1.668 1.668 0 00-2.52-1.337s-.583-2.416-2.489-1.928-3.418 5.043-3.15 6.192c.267 1.149 1.748.449 2.496-.63 0 0-.969 1.795-.323 2.267.945.692 1.938-.535 2.741-1.432 0 0-1.189 2.1.56 2.29 1.346.149 2.599-2.566 2.599-2.566l.874.622z"
+                            fill="#FFF"
+                        />
+                        <path
+                            d="M179.585 50.943h-.15c-.718 0-1.302-.58-1.307-1.299-.653.575-1.41.937-2.19.362a1.21 1.21 0 01-.354-1.235c-.34.213-.744.3-1.142.244a1.04 1.04 0 01-.827-.873c-.291-1.33 1.276-6.098 3.45-6.657a2.026 2.026 0 011.623.205c.56.384.984.934 1.213 1.573a1.987 1.987 0 011.575.095 2.36 2.36 0 011.008 1.424c.343.099.637.32.827.621.701 1.094.095 3.085 0 3.313a.37.37 0 01-.244.252.402.402 0 01-.355-.048l-.535-.37c-.378.803-1.386 2.393-2.592 2.393zm-.669-3.085a.37.37 0 01.236.079c.159.12.205.337.11.511-.173.307-.472 1.063-.283 1.432 0 .055.11.229.543.276.788.086 1.828-1.574 2.198-2.36a.402.402 0 01.252-.213.378.378 0 01.33.055l.442.307a3.31 3.31 0 00-.07-2.132.717.717 0 00-.544-.315.401.401 0 01-.34-.339 1.723 1.723 0 00-.676-1.125 1.49 1.49 0 00-1.3.095.402.402 0 01-.33 0 .41.41 0 01-.214-.26 2.77 2.77 0 00-.992-1.526 1.277 1.277 0 00-1.016-.11c-1.678.424-3.088 4.72-2.875 5.72.055.244.157.267.213.275a2.166 2.166 0 001.575-1.046c.12-.17.35-.219.528-.11a.401.401 0 01.141.519c-.401.787-.622 1.636-.44 1.77.55.401 1.204-.244 2.055-1.188l.166-.189a.402.402 0 01.291-.126z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M174.568 38.85s-1.079-1.873-2.213-.229c-1.134 1.644-2.497 4.72-.607 5.051 1.89.33 3.308-3.194 2.82-4.823z"
+                            fill="#FFF"
+                        />
+                        <path
+                            d="M171.992 44.074a2.019 2.019 0 01-.307 0 1.418 1.418 0 01-1.134-.85c-.551-1.314.67-3.651 1.48-4.815a1.63 1.63 0 011.371-.85c.867.04 1.442.999 1.505 1.101l.04.087a5.504 5.504 0 01-1.222 4.445 2.364 2.364 0 01-1.733.882zm1.355-5.752c-.26 0-.512.275-.67.504-.952 1.376-1.732 3.273-1.401 4.067.095.215.3.36.535.378a1.52 1.52 0 001.324-.622 4.765 4.765 0 001.063-3.658c-.126-.205-.48-.653-.835-.669h-.016z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M169.015 56.143s-1.244-5.94-1.363-7.002c-.196-1.802.229-4.619 1.812-5.06 1.583-.44 5.207-.857 4.608.732-.544 1.448-3.718 1.723-3.718 1.723s2.19 2.203 1.859 4.399"
+                            fill="#FFF"
+                        />
+                        <path
+                            d="M169.015 56.537a.394.394 0 01-.386-.307c-.047-.244-1.244-5.98-1.362-7.081-.205-1.842.189-4.95 2.095-5.508.992-.275 4.025-.842 4.915.126.274.305.34.745.165 1.117-.449 1.212-2.26 1.692-3.308 1.88.662.835 1.725 2.48 1.465 4.163a.398.398 0 01-.788-.118c.3-1.975-1.725-4.044-1.74-4.06a.409.409 0 01-.095-.417.394.394 0 01.339-.252c.787-.07 3.009-.472 3.379-1.471.086-.213 0-.276 0-.307-.418-.464-2.6-.323-4.128.094-1.283.354-1.709 2.92-1.52 4.635.118 1.03 1.347 6.916 1.355 6.97a.385.385 0 01-.307.465l-.079.07z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M192.275 33.617s-13.123-1.298-18.479 1.802a34.39 34.39 0 00-3.86 2.675c-2.906 2.274-5.639 4.854-7.766 6.294-4.221 2.833-8.301 3.407-11.082-1.101-2.363-3.848-.984-5.673 2.213-9.096 1.576-1.692 4.915-4.485 7.79-7.577a24.322 24.322 0 004.254-5.822c2.426-5.099 1.74-12.912 1.74-12.912l25.19 25.737z"
+                            fill="#FFF"
+                        />
+                        <path
+                            d="M156.27 46.977a5.533 5.533 0 01-1.26-.134 6.725 6.725 0 01-4.253-3.336c-2.544-4.123-.976-6.113 2.253-9.575.677-.724 1.654-1.637 2.788-2.684 1.576-1.463 3.348-3.147 5.01-4.894a23.692 23.692 0 004.182-5.72c2.363-4.957 1.71-12.668 1.702-12.746a.401.401 0 01.228-.394.386.386 0 01.449.087l25.205 25.76a.393.393 0 01.07.441.402.402 0 01-.393.228c-.134 0-13.06-1.25-18.242 1.763a33.76 33.76 0 00-3.812 2.643c-1.19.937-2.363 1.912-3.458 2.849a53.354 53.354 0 01-4.332 3.446c-2.284 1.51-4.332 2.266-6.136 2.266zM167.543 8.88c.095 2.36.087 8.01-1.835 12.094a24.739 24.739 0 01-4.324 5.916c-1.678 1.802-3.466 3.462-5.05 4.934-1.118 1.038-2.094 1.943-2.748 2.644-3.15 3.344-4.41 4.98-2.158 8.655a6.017 6.017 0 003.749 2.99c1.875.409 4.151-.276 6.774-2.03a50.708 50.708 0 004.261-3.384c1.126-.944 2.284-1.928 3.482-2.864a34.847 34.847 0 013.938-2.714c4.56-2.652 14.39-2.196 17.651-1.952L167.542 8.88z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M170.638 37.567c-2.907 2.274-6.42 5.287-8.468 6.814-4.678 3.493-7.64 3.863-10.696-.457-3.15-4.438-1.37-6.294 1.827-9.725 1.576-1.692 4.915-4.485 7.79-7.577 0 0 .181 3.218 3.387 7.144a11.17 11.17 0 006.16 3.8z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M156.24 47.473a4.374 4.374 0 01-.725-.055c-1.576-.244-2.962-1.283-4.364-3.266-3.316-4.681-1.378-6.75 1.82-10.181.716-.787 1.693-1.676 2.827-2.723 1.576-1.463 3.348-3.147 5.01-4.894a.378.378 0 01.417-.102c.148.053.25.19.26.346 0 0 .229 3.148 3.3 6.916a10.871 10.871 0 005.924 3.667.41.41 0 01.315.291.425.425 0 01-.142.41c-1.34 1.046-2.796 2.242-4.214 3.398-1.623 1.33-3.15 2.59-4.261 3.415-2.513 1.896-4.482 2.778-6.168 2.778zm4.584-19.993c-1.505 1.573-3.08 3.037-4.49 4.343-1.119 1.038-2.095 1.943-2.749 2.644-3.15 3.39-4.726 5.067-1.788 9.229 1.268 1.794 2.489 2.73 3.836 2.943 1.63.251 3.568-.535 6.301-2.573 1.095-.787 2.623-2.07 4.238-3.391 1.197-.984 2.426-1.991 3.591-2.92a11.705 11.705 0 01-5.592-3.745 15.73 15.73 0 01-3.347-6.53z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M195.914 32.507c-2.938 2.636-9.452.441-15.793-4.917a45.642 45.642 0 01-8.412-9.371c-4.726-6.94-6.199-13.66-3.237-16.311 3.654-3.281 12.775.905 20.36 9.355 7.586 8.45 10.752 17.955 7.082 21.244z"
+                            fill="#7C5DFA"
+                        />
+                        <path
+                            d="M171.772.798c4.466 0 11.208 3.982 17.037 10.473 7.593 8.45 10.775 17.955 7.089 21.244a4.832 4.832 0 01-3.324 1.118c-3.332 0-7.924-2.211-12.469-6.035a45.418 45.418 0 01-4.568-4.438 45.642 45.642 0 01-3.828-4.941c-4.726-6.94-6.199-13.66-3.237-16.311a4.84 4.84 0 013.3-1.11m0-.786a5.58 5.58 0 00-3.852 1.321c-3.205 2.872-1.953 9.836 3.112 17.31a47.8 47.8 0 008.553 9.553c4.726 3.934 9.452 6.215 12.973 6.215a5.58 5.58 0 003.883-1.353c2.025-1.81 2.363-5.193.89-9.52a38.548 38.548 0 00-7.931-12.833C183.524 4.22 176.624.012 171.772.012z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M185.359 21.043l-5.238 6.547a45.642 45.642 0 01-8.412-9.371l7.577-4.65 6.073 7.474z"
+                            fill="#FFF"
+                        />
+                        <ellipse
+                            fill="#231F20"
+                            transform="rotate(-39.4 182.308 17.334)"
+                            cx="182.308"
+                            cy="17.334"
+                            rx="2.244"
+                            ry="4.817"
+                        />
+                        <path
+                            d="M184.556 21.689c-1.237 0-2.907-1-4.301-2.7a8.786 8.786 0 01-1.733-3.194c-.291-1.133-.126-2.022.465-2.502.59-.48 1.489-.472 2.544 0a8.662 8.662 0 012.78 2.353c1.86 2.258 2.419 4.76 1.276 5.704-.296.225-.66.344-1.031.339zm-4.498-7.963a.883.883 0 00-.575.173c-.33.275-.402.897-.197 1.7a7.866 7.866 0 001.575 2.895c1.655 2.014 3.56 2.801 4.238 2.243.677-.56.276-2.581-1.378-4.596a7.993 7.993 0 00-2.56-2.108 2.759 2.759 0 00-1.103-.307zM178.916 48.645a.37.37 0 01-.197-.055.386.386 0 01-.142-.535 9.515 9.515 0 011.796-2.148.394.394 0 11.504.598 9.184 9.184 0 00-1.575 1.951.402.402 0 01-.386.19zM176.498 47.811h-.063a.394.394 0 01-.307-.464 5.012 5.012 0 011.882-2.801.394.394 0 01.473.63 4.272 4.272 0 00-1.576 2.32.402.402 0 01-.41.315zM182.074 48.37a.393.393 0 01-.315-.622l.536-.787a.41.41 0 01.551-.094.401.401 0 01.095.55l-.544.787a.378.378 0 01-.323.166z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M157.216 46.678a.387.387 0 01-.236-.079c-3.482-2.643-5.64-5.854-6.065-9.032v-.087a.393.393 0 01.173-.448.386.386 0 01.544.125c.046.096.073.2.078.307.402 2.975 2.442 5.996 5.766 8.506a.393.393 0 01-.244.708h-.016zM155.27 34.12a.378.378 0 01-.283-.133.385.385 0 010-.551l4.238-3.934a.378.378 0 01.55 0 .385.385 0 010 .55l-4.237 3.935a.378.378 0 01-.267.133zM155.853 35.065a.426.426 0 01-.291-.126.401.401 0 010-.559l4.238-3.934a.394.394 0 01.667.28.393.393 0 01-.116.279l-4.23 3.934a.418.418 0 01-.268.126zM156.523 36.15a.362.362 0 01-.284-.126.385.385 0 010-.55l4.238-3.934a.378.378 0 01.551 0 .385.385 0 010 .55l-4.237 3.934a.37.37 0 01-.268.126z"
+                            fill="#FFF"
+                        />
+                        <path
+                            d="M86.422 96.838l-1.977 43.606s26.977.322 42.494-6.153a101.881 101.881 0 014.6-12.188c2.276-4.595 12.05-11.37 8.979-28.963 0 0 23.03 14.556 29.521 7.868 6.49-6.688 9.736-37.925 9.736-37.925l-13.485-4.327-6.128 19.261S121.299 59.236 112.24 56.86c-3.363-.88-13.816 4.674-25.796 12.361"
+                            fill="#7C5DFA"
+                        />
+                        <path
+                            d="M111.406 69.936s6.246 19.388 7.317 27.594c1.072 8.207 1.23 21.772 1.23 21.772s9.884-30.246 8.663-48.925l-17.21-.44z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M119.952 119.656h-.047a.346.346 0 01-.3-.339c0-.133-.173-13.643-1.228-21.732-1.056-8.088-7.239-27.342-7.302-27.538a.354.354 0 01.339-.457l17.202.44c.181 0 .33.143.339.323 1.22 18.53-8.57 48.783-8.664 49.059a.347.347 0 01-.339.244zm-8.066-49.358c.954 3.006 6.215 19.75 7.184 27.185.788 6.185 1.095 15.508 1.19 19.67 2.15-7.08 8.94-30.82 8.025-46.422l-16.399-.433zM79.94 163.592h74.44l-27.418-29.301-42.494 6.153s-5.088 11.503-4.608 19.576c.032.952.056 2.171.08 3.572z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M154.38 163.985H79.94a.387.387 0 01-.395-.385c0-1.322-.047-2.589-.102-3.557-.48-8.08 4.427-19.285 4.64-19.765a.41.41 0 01.299-.228l42.533-6.153a.41.41 0 01.347.126l27.418 29.301a.393.393 0 01-.292.661h-.008zm-74.04-.787h73.127l-26.654-28.49-42.077 6.09c-.67 1.573-4.93 11.889-4.49 19.198.04.89.095 2.014.095 3.202z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M126.474 59.511l2.134 10.866s.056 6.295-6.923 5.689a13.92 13.92 0 01-10.279-6.13l-2.828-18.978a34.302 34.302 0 00-.835-10.3c8.027-1.66 13.21-8.922 13.21-8.922s3.552 3.62 5.001 4.721a16.535 16.535 0 013.15 2.81c1.52 1.88 1.97 4.508 2.364 6.829.495 2.755.74 5.549.732 8.348-.126 5.311-5.726 5.067-5.726 5.067z"
+                            fill="#FFF"
+                        />
+                        <path
+                            d="M122.59 76.467h-.937a14.18 14.18 0 01-10.554-6.294.456.456 0 01-.047-.134s-2.82-18.884-2.828-18.978a4.802 4.802 0 010-.952 31.897 31.897 0 00-.835-9.332.322.322 0 01.047-.268.362.362 0 01.229-.157c7.79-1.574 12.933-8.71 12.98-8.781a.37.37 0 01.268-.15c.107-.01.212.031.284.11.039 0 3.552 3.612 4.97 4.674a16.92 16.92 0 013.15 2.864c1.576 1.983 2.056 4.721 2.434 6.995.501 2.776.75 5.591.74 8.411a5.182 5.182 0 01-1.575 3.872 6.62 6.62 0 01-4.072 1.573l2.056 10.441a6.621 6.621 0 01-1.922 4.627 6.306 6.306 0 01-4.387 1.48zm-10.837-6.656a13.692 13.692 0 009.963 5.9c2.08.174 3.702-.259 4.829-1.298a5.992 5.992 0 001.701-4.036l-2.127-10.795a.393.393 0 01.08-.307.339.339 0 01.29-.126 6.156 6.156 0 004.002-1.377 4.537 4.537 0 001.315-3.352 45.872 45.872 0 00-.732-8.285c-.386-2.258-.788-4.831-2.284-6.656a16.432 16.432 0 00-3.049-2.746c-1.228-.921-3.867-3.549-4.726-4.446-1.126 1.456-5.844 7.081-12.815 8.655.704 3.002.97 6.09.788 9.167a4.889 4.889 0 000 .786s-.032.197 2.765 18.916z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M127.427 64.861s-4.38 1.031-7.782-1.369c-3.403-2.4-3.545-5.444-3.545-5.444s5.388 2.29 10.374 1.463l.953 5.35z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M125.135 65.436a9.845 9.845 0 01-5.687-1.66c-3.505-2.471-3.663-5.54-3.663-5.713a.322.322 0 01.15-.299.331.331 0 01.33-.039c.056 0 5.365 2.242 10.185 1.44a.347.347 0 01.394.283l.961 5.35a.346.346 0 01-.268.402 10.81 10.81 0 01-2.402.236zm-8.562-6.845a7.916 7.916 0 003.3 4.618c2.718 1.92 6.136 1.574 7.184 1.377l-.835-4.682c-4.057.56-8.184-.763-9.649-1.313z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M120.661 40.981s4.497 4.564 3.836 5.626c-.662 1.062-3.104.913-3.104.913"
+                            fill="#FFF"
+                        />
+                        <path
+                            d="M121.772 47.874h-.402a.347.347 0 01-.323-.37.354.354 0 01.37-.33c.622 0 2.363 0 2.788-.74.268-.44-1.496-2.872-3.788-5.2a.347.347 0 11.496-.489c1.41 1.432 4.608 4.887 3.883 6.059-.598.952-2.245 1.07-3.024 1.07z"
+                            fill="#231F20"
+                        />
+                        <ellipse
+                            fill="#231F20"
+                            transform="rotate(-13.01 126.624 42.344)"
+                            cx="126.624"
+                            cy="42.344"
+                            rx="1"
+                            ry="1.149"
+                        />
+                        <ellipse
+                            fill="#231F20"
+                            transform="rotate(-13.01 116.589 43.363)"
+                            cx="116.589"
+                            cy="43.363"
+                            rx="1"
+                            ry="1.149"
+                        />
+                        <path
+                            d="M111.154 44.695c-1.095-1.77-3.655-2.014-5.443-1.298a3.148 3.148 0 00-2.087 2.848c-.014.773.125 1.54.41 2.258a11.87 11.87 0 002.11 3.69 4.253 4.253 0 001.647 1.323c.839.296 1.774.053 2.363-.614"
+                            fill="#FFF"
+                        />
+                        <path
+                            d="M108.531 54.02c-.297-.001-.59-.057-.866-.166a4.725 4.725 0 01-1.796-1.424 12.35 12.35 0 01-2.174-3.8 6.289 6.289 0 01-.433-2.408 3.51 3.51 0 012.315-3.148c1.97-.786 4.726-.472 5.884 1.44a.362.362 0 01-.614.378c-.985-1.574-3.308-1.833-5.002-1.157a2.817 2.817 0 00-1.874 2.526 5.59 5.59 0 00.385 2.124 11.563 11.563 0 002.048 3.58c.393.523.91.94 1.505 1.212.712.25 1.505.041 2-.527a.363.363 0 11.512.504c-.489.528-1.17.84-1.89.865z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M125.072 48.936l-5.27.512s1.095 8.135 3.466 7.915c2.37-.22 2.379-6.098 1.804-8.427z"
+                            fill="#FFF"
+                        />
+                        <path
+                            d="M123.166 57.717c-2.481 0-3.529-6.83-3.71-8.222a.307.307 0 01.07-.26.315.315 0 01.237-.134l5.277-.511a.347.347 0 01.37.26c.45 1.81.6 6.145-.708 7.978-.31.49-.826.814-1.402.881l-.134.008zm-2.97-7.955c.441 2.975 1.576 7.255 2.97 7.255h.07a1.26 1.26 0 00.899-.59c1.031-1.456 1.079-5.099.669-7.082l-4.608.417zM138.257 112.024a.355.355 0 01-.3-.543c.056-.103 5.845-9.828.308-30.08a.37.37 0 01.252-.441.355.355 0 01.441.252c5.632 20.583-.142 30.245-.394 30.686a.347.347 0 01-.307.126z"
+                            fill="#231F20"
+                        />
+                        <path
+                            fill="#FFF"
+                            d="M31.136 83.879l85.035 61.285 85.036-61.285v100.54H31.136z"
+                        />
+                        <path
+                            d="M201.207 184.812H31.136a.394.394 0 01-.394-.393V83.879a.378.378 0 01.213-.346.386.386 0 01.41 0l84.806 61.112 84.8-61.112a.386.386 0 01.409 0c.134.063.22.198.22.346v100.54a.401.401 0 01-.393.393zm-169.7-.786h169.345V84.65L116.4 145.48a.394.394 0 01-.465 0L31.506 84.65v99.376z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M31.136 184.812a.395.395 0 01-.347-.204.401.401 0 01.15-.535l77.805-44.463a.394.394 0 01.535.149.385.385 0 01-.141.535l-77.813 44.463a.402.402 0 01-.189.055zM201.207 184.812a.434.434 0 01-.197-.055l-77.805-44.463a.393.393 0 01-.15-.535.402.402 0 01.536-.15l77.805 44.464a.393.393 0 01.15.535.386.386 0 01-.34.204z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M71.677 138.303s3.442 8.954 4.253 9.442c.811.488 10.397 4.406 9.712 6.24-.685 1.833-5.002-.402-5.002-.402s6.215 7.184 4.576 8.451c-1.638 1.267-7.151 3.84-9.365 2.793-2.213-1.046-13.248-14.706-15.587-20.638"
+                            fill="#FFF"
+                        />
+                        <path
+                            d="M77.19 165.488a3.532 3.532 0 01-1.496-.283c-2.363-1.125-13.47-14.95-15.793-20.85a.396.396 0 11.74-.284c2.364 6.05 13.47 19.513 15.384 20.457 1.914.944 7.293-1.448 8.963-2.746a.37.37 0 00.118-.26c.15-1.573-3.387-6.034-4.726-7.624a.393.393 0 01.48-.606c.788.41 3.041 1.385 4.05 1a.646.646 0 00.401-.394c.189-.511-1.055-1.794-8.073-5.043-.851-.394-1.3-.606-1.481-.724-.851-.543-3.592-7.53-4.411-9.654a.393.393 0 11.732-.284c1.639 4.257 3.624 8.891 4.112 9.285.134.086.709.354 1.378.66 5.656 2.613 9.043 4.54 8.483 6.036-.146.397-.46.71-.858.857a5.109 5.109 0 01-3.15-.362c1.52 1.897 4.032 5.303 3.874 6.924a1.133 1.133 0 01-.417.787c-1.3.984-5.49 3.108-8.31 3.108z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M86.445 69.252C66.187 82.25 41.55 101.378 39.47 105.847c-3.323 7.081 21.535 42.693 21.535 42.693l13.461-8.53-12.303-26.696 24.26-16.484"
+                            fill="#7C5DFA"
+                        />
+                        <path
+                            d="M70.889 107.806a.355.355 0 01-.3-.15.37.37 0 01.087-.503l27.237-19.049a.355.355 0 01.505.087.362.362 0 01-.095.503l-27.23 19.05a.332.332 0 01-.204.062zM189.234 161.79l33.01-20.182-27.969 11.637z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M189.234 162.183a.386.386 0 01-.275-.118.393.393 0 01-.063-.48l5.04-8.537a.37.37 0 01.182-.165l28-11.637a.394.394 0 01.355.7l-33.034 20.15a.465.465 0 01-.205.087zm5.31-8.655l-4.168 7.082 27.324-16.704-23.157 9.622z"
+                            fill="#231F20"
+                        />
+                        <path
+                            fill="#FFF"
+                            d="M194.275 153.245l10.791 5.516 17.179-17.153z"
+                        />
+                        <path
+                            d="M205.066 159.115a.323.323 0 01-.165-.04l-10.791-5.507a.346.346 0 01-.197-.33.378.378 0 01.22-.323l27.986-11.637a.363.363 0 01.433.133.346.346 0 01-.047.449l-17.179 17.153a.331.331 0 01-.26.102zm-9.94-5.83l9.87 5.035L220.7 142.64l-25.575 10.645z"
+                            fill="#231F20"
+                        />
+                        <path fill="#FFF" d="M222.245 141.608l-35.54 8.482-8.08-4.619z" />
+                        <path
+                            d="M186.674 150.444a.339.339 0 01-.173-.047l-8.09-4.61a.362.362 0 01.15-.67l43.62-3.863a.362.362 0 01.387.299.362.362 0 01-.268.41l-35.539 8.473-.087.008zm-6.907-4.72l6.963 3.973 30.663-7.27-37.626 3.296z"
+                            fill="#231F20"
+                        />
+                        <path
+                            fill="#FFF"
+                            d="M186.706 150.09l2.528 11.7 5.041-8.545 27.97-11.637z"
+                        />
+                        <path
+                            d="M189.234 162.152h-.055a.362.362 0 01-.3-.275l-2.52-11.708a.354.354 0 01.268-.425l35.54-8.474a.356.356 0 01.22.676l-27.836 11.59-4.978 8.435a.355.355 0 01-.339.181zm-2.103-11.802l2.26 10.457 4.57-7.75a.402.402 0 01.172-.15l23.835-9.914-30.837 7.357z"
+                            fill="#231F20"
+                        />
+                        <path
+                            d="M181.948 163.419a.394.394 0 01-.204-.724l1.315-.842a.396.396 0 01.433.66l-1.33.85a.426.426 0 01-.214.056zM62.713 199.47a.401.401 0 01-.394-.392c0-.215.172-.39.386-.394l1.875-.04 1.268-.039a.41.41 0 01.41.378.393.393 0 01-.386.41l-1.268.039-1.883.039h-.008zm-3.15 0c-1.072 0-2.127-.046-3.151-.086a.401.401 0 01-.378-.409.393.393 0 01.41-.377c1.023.039 2.07.07 3.15.086a.394.394 0 01.394.394.401.401 0 01-.418.393h-.007zm9.451-.212a.386.386 0 01-.386-.37.393.393 0 01.37-.416 143.63 143.63 0 003.151-.181.386.386 0 01.418.37.385.385 0 01-.363.417c-1.047.07-2.095.133-3.15.18h-.04zm-15.713-.055a95.648 95.648 0 01-3.151-.283.394.394 0 01-.347-.433.378.378 0 01.434-.346c1.016.102 2.055.197 3.15.276a.394.394 0 110 .786h-.086zm21.999-.354a.394.394 0 110-.786c1.047-.087 2.087-.181 3.15-.284a.41.41 0 01.434.354.393.393 0 01-.355.433c-1.05.1-2.1.194-3.15.283h-.08zm-28.261-.322h-.055c-1.056-.158-2.103-.33-3.151-.52a.4.4 0 11.15-.786c1 .188 2.04.362 3.087.511a.394.394 0 01.331.449.402.402 0 01-.378.346h.016zm34.53-.284a.395.395 0 11-.047-.786c1.05-.116 2.1-.242 3.151-.378a.397.397 0 11.102.787l-3.15.37-.055.007zm6.247-.786a.402.402 0 01-.394-.339.393.393 0 01.339-.448l3.103-.464a.386.386 0 01.449.322.393.393 0 01-.323.457l-3.15.464-.024.008zm-46.952-.079h-.095a72.266 72.266 0 01-3.064-.787.401.401 0 01-.268-.488.394.394 0 01.489-.267c.976.275 1.992.543 3.024.787a.386.386 0 01.292.472.394.394 0 01-.394.275l.016.008zm53.166-.905a.386.386 0 01-.386-.322.401.401 0 01.323-.457l3.096-.55a.386.386 0 01.457.314.393.393 0 01-.315.456l-3.104.56h-.07zm-59.223-.857a.355.355 0 01-.127 0 52.574 52.574 0 01-2.961-1.141.394.394 0 01.307-.724c.937.393 1.922.787 2.914 1.125a.4.4 0 11-.133.787v-.047zm65.375-.3a.402.402 0 01-.386-.314.393.393 0 01.307-.464l3.08-.646a.394.394 0 01.465.307.385.385 0 01-.3.464l-3.087.646-.08.008zm6.16-1.321a.396.396 0 01-.087-.787c1.024-.236 2.04-.472 3.056-.724.21-.05.422.08.472.291a.393.393 0 01-.291.48c-1.016.244-2.032.488-3.064.716l-.087.024zm-77.34-.842a.418.418 0 01-.182-.047 34.933 34.933 0 01-2.764-1.574.393.393 0 01-.126-.543.386.386 0 01.543-.126 40.365 40.365 0 002.71 1.519.404.404 0 11-.182.786v-.015zm83.491-.645a.397.397 0 11-.095-.787l3.04-.787a.407.407 0 01.206.787c-1.009.273-2.025.535-3.049.787a.213.213 0 01-.118 0h.016zm6.08-1.645a.386.386 0 01-.377-.283.401.401 0 01.275-.488l3.017-.873a.394.394 0 01.488.26.385.385 0 01-.267.487c-1 .3-2.009.598-3.025.881a.268.268 0 01-.126.016h.016zm-94.888-1.015a.386.386 0 01-.236-.079 32.502 32.502 0 01-2.458-2.014.393.393 0 01.551-.558c.768.695 1.57 1.351 2.403 1.967.174.13.209.377.078.55a.402.402 0 01-.338.134zm100.906-.787a.378.378 0 01-.37-.275.393.393 0 01.26-.496c1.008-.307 2-.63 2.993-.944a.386.386 0 01.496.252.393.393 0 01-.252.495c-.992.323-1.992.638-3.008.953a.307.307 0 01-.119.023v-.008zm5.994-1.943a.399.399 0 11-.126-.787c1-.338 1.993-.677 2.97-1.023.206-.068.43.04.504.244a.385.385 0 01-.236.496c-.985.346-1.977.692-2.986 1.03l-.126.04zm-111.54-1.495a.41.41 0 01-.299-.134 28.33 28.33 0 01-1.985-2.486.393.393 0 11.638-.457 26.828 26.828 0 001.938 2.416.401.401 0 010 .559.394.394 0 01-.291.11v-.008zm117.48-.582a.41.41 0 01-.37-.26.393.393 0 01.235-.504c1-.362 1.977-.731 2.954-1.093a.394.394 0 01.276.731c-.977.37-1.961.732-2.962 1.094a.26.26 0 01-.134.04v-.008zm5.891-2.227a.4.4 0 01-.142-.787l2.922-1.157c.2-.081.429.014.512.213a.401.401 0 01-.212.52l-2.93 1.195a.497.497 0 01-.15.024v-.008zm-127.01-2.298a.386.386 0 01-.346-.204 27.471 27.471 0 01-1.378-2.872.395.395 0 11.732-.3c.387.96.834 1.892 1.34 2.794.1.192.03.428-.158.535a.418.418 0 01-.19.047zm132.854-.055a.378.378 0 01-.362-.244.393.393 0 01.205-.511c.984-.41 1.953-.787 2.898-1.227a.395.395 0 01.316.723l-2.907 1.228a.339.339 0 01-.15.031zm5.782-2.494a.41.41 0 01-.363-.228.385.385 0 01.205-.52c.977-.44 1.93-.865 2.86-1.298a.395.395 0 01.33.716c-.937.433-1.89.866-2.875 1.299a.355.355 0 01-.157.031zm5.726-2.628a.378.378 0 01-.354-.22.393.393 0 01.18-.527c.977-.457 1.923-.913 2.836-1.362a.395.395 0 01.347.709l-2.844 1.369a.41.41 0 01-.165.031zm-146.74-.653a.394.394 0 01-.378-.291 28.685 28.685 0 01-.693-3.108.398.398 0 01.787-.126c.158 1.03.373 2.048.646 3.053a.393.393 0 01-.268.488l-.094-.016zm152.38-2.1a.386.386 0 01-.355-.221.393.393 0 01.174-.527c.968-.488 1.906-.96 2.804-1.432a.395.395 0 01.362.7c-.898.464-1.835.944-2.812 1.432a.363.363 0 01-.173.047zm5.592-2.904a.386.386 0 01-.347-.205.401.401 0 01.158-.535c.977-.52 1.898-1.023 2.765-1.495a.394.394 0 11.378.685c-.867.48-1.788.983-2.773 1.51a.363.363 0 01-.157.04H171zm-159.028-1.157a.394.394 0 01-.393-.37c0-.613-.048-1.243-.048-1.864v-1.314a.385.385 0 01.41-.378.394.394 0 01.378.41v1.282c0 .606 0 1.22.047 1.817a.393.393 0 01-.37.417h-.024zm164.542-1.888a.394.394 0 01-.197-.732c1.016-.582 1.93-1.11 2.717-1.574a.394.394 0 01.544.134c.11.188.05.428-.134.543a91.573 91.573 0 01-2.733 1.574.426.426 0 01-.174.055h-.023zm-164.266-4.383h-.055a.401.401 0 01-.339-.448c.142-1.031.323-2.085.56-3.148.05-.21.26-.343.472-.299a.394.394 0 01.3.473 34.676 34.676 0 00-.552 3.06.394.394 0 01-.386.362zm1.37-6.13a.307.307 0 01-.118 0 .394.394 0 01-.26-.495c.308-.991.67-1.998 1.072-2.99a.394.394 0 01.724.3c-.386.967-.74 1.95-1.047 2.926a.378.378 0 01-.37.26zm2.364-5.822a.363.363 0 01-.229-.007.394.394 0 01-.181-.528c.465-.936.977-1.872 1.52-2.785a.386.386 0 01.536-.142c.187.113.25.354.142.543-.536.89-1.032 1.81-1.49 2.738a.402.402 0 01-.322.174l.024.007zm3.205-5.413a.41.41 0 01-.283-.102.386.386 0 01-.095-.543 42.74 42.74 0 011.906-2.542.394.394 0 01.615.496 36.813 36.813 0 00-1.875 2.494.402.402 0 01-.291.197h.023zm3.939-4.878a.386.386 0 01-.284-.661 33.592 33.592 0 012.276-2.219.386.386 0 01.552 0 .393.393 0 010 .559 32.391 32.391 0 00-2.23 2.18.402.402 0 01-.314.14zM27.836 135.77a.386.386 0 01-.307-.15.401.401 0 01.07-.558c.426-.315.851-.63 1.284-.937a.394.394 0 11.45.646c-.426.299-.843.605-1.253.92a.394.394 0 01-.244.079zM204.475 88.757a.355.355 0 01-.267-.11.385.385 0 010-.55s.386-.41 1.055-1.15a.394.394 0 11.583.52c-.677.787-1.071 1.172-1.071 1.172a.386.386 0 01-.3.118zM207.626 85.209a.41.41 0 01-.26-.095.385.385 0 010-.55c.607-.724 1.284-1.574 2.009-2.448a.394.394 0 01.7.2.385.385 0 01-.086.288 97.323 97.323 0 01-2.024 2.463.386.386 0 01-.339.142zm3.938-4.957a.45.45 0 01-.236-.079.401.401 0 01-.07-.559 93.958 93.958 0 001.874-2.541.396.396 0 01.646.456c-.646.897-1.276 1.755-1.899 2.565a.386.386 0 01-.283.173l-.032-.015zm3.702-5.17a.386.386 0 01-.323-.613 91.041 91.041 0 001.725-2.652.402.402 0 01.544-.126c.182.118.235.36.118.543a94.643 94.643 0 01-1.733 2.675.394.394 0 01-.3.173h-.03zm3.371-5.382a.355.355 0 01-.196-.055.386.386 0 01-.142-.535 65.203 65.203 0 001.512-2.77.395.395 0 01.701.363 66.95 66.95 0 01-1.536 2.8.394.394 0 01-.307.197h-.032zm2.923-5.641a.41.41 0 01-.166 0 .386.386 0 01-.189-.52c.45-.983.867-1.959 1.229-2.903a.394.394 0 01.512-.228c.202.08.304.307.228.511-.378.96-.787 1.952-1.252 2.943a.41.41 0 01-.33.197h-.032zm2.292-5.917h-.118a.401.401 0 01-.26-.496 34.59 34.59 0 00.787-3.037.394.394 0 01.772.165 63.31 63.31 0 01-.787 3.148.402.402 0 01-.363.22h-.031zm1.323-6.2h-.04a.394.394 0 01-.354-.425c0-.33.063-.653.079-.984 0-.731.07-1.455.094-2.155a.385.385 0 01.402-.386.401.401 0 01.386.401c0 .716-.055 1.44-.094 2.188 0 .338-.048.668-.08 1.007a.402.402 0 01-.393.354zm.181-6.342a.394.394 0 01-.394-.385c0-1.07-.078-2.125-.157-3.148a.385.385 0 01.362-.417.394.394 0 01.425.362c.08 1.03.134 2.101.158 3.187 0 .215-.17.393-.386.401h-.008zm-.512-6.294a.402.402 0 01-.394-.339 36.894 36.894 0 00-.575-3.084.403.403 0 11.788-.173c.228 1.007.425 2.07.583 3.147a.401.401 0 01-.33.449h-.072zm-1.425-6.17a.394.394 0 01-.379-.267 28.307 28.307 0 00-1.126-2.91.394.394 0 11.717-.331c.441.977.828 1.977 1.158 2.997a.393.393 0 01-.244.496.37.37 0 01-.126-.008v.024zm-2.631-5.75a.402.402 0 01-.339-.182 21.968 21.968 0 00-1.812-2.533.393.393 0 01.056-.551.386.386 0 01.55.047 23.22 23.22 0 011.876 2.597.385.385 0 01-.331.598v.023zm-4.049-4.856a.426.426 0 01-.26-.094 20.532 20.532 0 00-2.473-1.888.393.393 0 11.425-.661 21.556 21.556 0 012.576 1.959.401.401 0 010 .559.402.402 0 01-.268.102v.023zm-5.3-3.438a.426.426 0 01-.166 0c-.949-.431-1.925-.8-2.922-1.101a.41.41 0 11.228-.787c1.033.309 2.043.687 3.025 1.133a.386.386 0 01.189.519.386.386 0 01-.355.212v.024zm-6.066-1.849h-.07a26.681 26.681 0 00-3.112-.354.394.394 0 01-.37-.417.37.37 0 01.417-.37c1.072.06 2.14.181 3.198.362.214.04.357.242.323.456a.386.386 0 01-.386.3v.023zM197.442 17.015a.395.395 0 01-.063-.787 14.303 14.303 0 011.575-.157.394.394 0 11.047.786c-.953.056-1.52.142-1.52.142l-.04.016z"
+                            fill="#D1D3D4"
+                        />
+                        <path
+                            d="M2.333 48.385l22.852-18.658a2.578 2.578 0 013.626.365l9.231 11.282a2.578 2.578 0 01-.365 3.63L14.825 63.662a2.578 2.578 0 01-3.626-.365L1.968 52.015a2.578 2.578 0 01.365-3.63z"
+                            fill="#FFF"
+                        />
+                        <path
+                            d="M13.122 64.87h-.244a2.364 2.364 0 01-1.575-.866L1.3 51.808a2.359 2.359 0 01.338-3.36l23.748-19.395a2.365 2.365 0 013.363.339l9.972 12.187a2.359 2.359 0 01-.339 3.36L14.635 64.326c-.425.352-.96.544-1.513.543zm13.776-35.565a1.647 1.647 0 00-1.016.354L2.135 49.054c-.324.27-.525.659-.56 1.078-.04.425.088.848.355 1.18l9.971 12.18c.267.332.656.542 1.08.583.422.04.843-.088 1.173-.354l23.732-19.396a1.573 1.573 0 00.229-2.25l-9.972-12.164a1.576 1.576 0 00-1.245-.606z"
+                            fill="#D1D3D4"
+                        />
+                        <path
+                            d="M1.181 50.368a.394.394 0 01-.393-.393.385.385 0 01.354-.425l19.794-1.715L26.56 28.8a.386.386 0 01.488-.267.393.393 0 01.26.488l-5.687 19.293a.386.386 0 01-.339.283l-20.061 1.77h-.04z"
+                            fill="#D1D3D4"
+                        />
+                        <path
+                            d="M22.054 45.773a.394.394 0 110-.787L38.8 43.074a.402.402 0 01.433.347.393.393 0 01-.346.44l-16.762 1.912h-.07zM13.099 64.87a.315.315 0 01-.118 0 .386.386 0 01-.252-.496l5.206-16.012a.412.412 0 01.788.244l-5.207 16.012a.402.402 0 01-.417.251z"
+                            fill="#D1D3D4"
+                        />
+                        <path
+                            d="M217.625 104.805l21.774 9.431a2.068 2.068 0 011.078 2.721l-.002.001-4.668 10.755a2.075 2.075 0 01-2.724 1.076l-21.775-9.43a2.068 2.068 0 01-1.075-2.723l4.667-10.754a2.075 2.075 0 012.725-1.077z"
+                            fill="#FFF"
+                        />
+                        <path
+                            d="M234.162 129.491a2.034 2.034 0 01-.788-.165l-22.621-9.804c-.49-.206-.876-.6-1.071-1.094-.201-.491-.201-1.042 0-1.534l5.033-11.606a2.01 2.01 0 012.63-1.038l22.63 9.796a2 2 0 011.063 1.101 1.973 1.973 0 010 1.527l-5.04 11.59a1.984 1.984 0 01-1.095 1.062 2.026 2.026 0 01-.74.165zm-17.557-24.62c-.483.002-.919.29-1.11.732l-5.073 11.645c-.258.608.024 1.31.63 1.574l22.63 9.804c.298.121.631.121.93 0 .298-.117.537-.35.66-.646l5.042-11.613a1.203 1.203 0 000-.92 1.173 1.173 0 00-.646-.67l-22.614-9.803a1.12 1.12 0 00-.449-.103z"
+                            fill="#D1D3D4"
+                        />
+                        <path
+                            d="M224.742 118.617a.378.378 0 01-.331-.18l-8.664-13.676a.394.394 0 11.67-.417l8.49 13.447 15.635-2.998a.386.386 0 01.465.307.401.401 0 01-.315.464l-15.895 3.053h-.055z"
+                            fill="#D1D3D4"
+                        />
+                        <path
+                            d="M234.619 129.42a.402.402 0 01-.33-.173l-7.547-11.212a.393.393 0 01.406-.603c.102.02.192.082.248.17l7.546 11.244c.12.183.07.428-.11.55a.418.418 0 01-.213.024zM210.107 118.822a.394.394 0 01-.394-.33.393.393 0 01.331-.45l13.39-2.187a.4.4 0 01.126.787l-13.39 2.188-.063-.008z"
+                            fill="#D1D3D4"
+                        />
+                        <path
+                            d="M65.276 12.726l16.47 5.454a1.51 1.51 0 01.96 1.91l-2.7 8.139a1.515 1.515 0 01-1.912.96l-16.47-5.454a1.51 1.51 0 01-.96-1.911l2.7-8.138a1.515 1.515 0 011.912-.96z"
+                            fill="#FFF"
+                        />
+                        <path
+                            d="M78.766 29.738c-.166 0-.331-.027-.489-.079l-17.116-5.665a1.575 1.575 0 01-.992-1.983l2.914-8.78a1.576 1.576 0 011.977-1l17.116 5.665a1.573 1.573 0 01.993 1.983l-2.907 8.78a1.575 1.575 0 01-1.496 1.079zm-.245-.826a.788.788 0 00.599 0 .787.787 0 00.394-.457l2.922-8.773a.786.786 0 000-.598.787.787 0 00-.449-.393l-17.116-5.665a.788.788 0 00-.598.047.787.787 0 00-.386.448l-2.915 8.781a.825.825 0 000 .598c.096.184.26.322.457.386l17.092 5.626z"
+                            fill="#D1D3D4"
+                        />
+                        <path
+                            d="M71.322 22.436a.402.402 0 01-.315-.157L63.8 12.884a.393.393 0 01.622-.48l7.018 9.182 11.122-3.147a.394.394 0 01.488.275.393.393 0 01-.275.48L71.4 22.452l-.079-.016z"
+                            fill="#D1D3D4"
+                        />
+                        <path
+                            d="M79.215 29.66a.378.378 0 01-.308-.15l-6.222-7.696a.394.394 0 11.614-.495l6.223 7.703a.393.393 0 01-.307.637zM60.65 23.522a.397.397 0 01-.095-.787l9.594-2.455a.394.394 0 01.48.283.385.385 0 01-.283.473l-9.594 2.462-.103.024z"
+                            fill="#D1D3D4"
+                        />
+                    </g>
+                </svg>
+            </SVGContainer>
+            <NothingHere>There is nothing here</NothingHere>
+            <Description>
+        Create a new invoice by clicking the{" "}
+                <span className="newInvoice">
+          New <span className="newInvoiceLargeOnly">Invoice</span>
+                </span>{" "}
+        button and get started
+            </Description>
+        </EmptyListContainer>
+    );
+}
+
+export default EmptyList;
