+++ src/components/form-components/FormDropDown.jsx
@@ -1,14 +1,14 @@
 // eslint-disable-next-line no-unused-vars
-import React, { useState, useCallback, useEffect } from "react";
+import React, { useState, useEffect } from "react";
 import styled from "styled-components";
 import PropTypes from "prop-types";
-import { SVG } from "../buttons/NewItemButton";
 import { useParams } from "react-router-dom";
 import { useSelector } from "react-redux";
+import { SVG } from "../buttons/NewItemButton";
 import { selectInvoiceById } from "../../features/invoices/invoicesSlice";
 
 const Main = styled.div.attrs({
-  tabIndex: 0,
+    tabIndex: 0,
 })`
   display: inline;
   z-index: 10;
@@ -152,73 +152,73 @@ const ItemButton = styled.button`
 const options = ["Net 1 Day", "Net 7 Days", "Net 14 Days", "Net 30 Days"];
 
 const arrowDown = (
-  <svg width="11" height="7" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M1 1l4.228 4.228L9.456 1"
-      stroke="#7C5DFA"
-      strokeWidth="2"
-      fill="none"
-      fillRule="evenodd"
-    />
-  </svg>
+    <svg width="11" height="7" xmlns="http://www.w3.org/2000/svg">
+        <path
+            d="M1 1l4.228 4.228L9.456 1"
+            stroke="#7C5DFA"
+            strokeWidth="2"
+            fill="none"
+            fillRule="evenodd"
+        />
+    </svg>
 );
 
 function FormDropDown({
-  selectedPaymentOption,
-  handleChangeSelectedOption,
-  isPaymentOpen,
-  handlePaymentClick,
+    selectedPaymentOption,
+    handleChangeSelectedOption,
+    isPaymentOpen,
+    handlePaymentClick,
 }) {
 
-  const { id } = useParams();
-  const invoice = useSelector(state => selectInvoiceById(state, id));
-  // const selectedPaymentOption = invoice?.paymentTerms || 1;
-  const [selected, setSelected] = useState("Net 1 Day");
+    const { id } = useParams();
+    const invoice = useSelector(state => selectInvoiceById(state, id));
+    // const selectedPaymentOption = invoice?.paymentTerms || 1;
+    const [selected, setSelected] = useState("Net 1 Day");
 
-  const onOptionClicked = (option) => (e) => {
-    e.preventDefault();
-    handlePaymentClick();
+    const onOptionClicked = (option) => (e) => {
+        e.preventDefault();
+        handlePaymentClick();
 
-    const num = Number(option.split(" ")[1]);
-    handleChangeSelectedOption(num);
-  };
+        const num = Number(option.split(" ")[1]);
+        handleChangeSelectedOption(num);
+    };
 
-  useEffect(() => {
-    if (selectedPaymentOption === 1) {
-      setSelected("Net 1 Day");
-    } else if (String(selectedPaymentOption).match(/\d+/)) {
-      setSelected(`Net ${selectedPaymentOption} Days`);
-    }
-  }, [selectedPaymentOption]);
+    useEffect(() => {
+        if (selectedPaymentOption === 1) {
+            setSelected("Net 1 Day");
+        } else if (String(selectedPaymentOption).match(/\d+/)) {
+            setSelected(`Net ${selectedPaymentOption} Days`);
+        }
+    }, [selectedPaymentOption]);
 
 
-  return (
-    <Main>
-      <DropDownHeader onClick={handlePaymentClick}
-        tabIndex={-1}
-      >
-        <h2>{selected}</h2>
-        <SVG>{arrowDown}</SVG>
-      </DropDownHeader>
+    return (
+        <Main>
+            <DropDownHeader onClick={handlePaymentClick}
+                tabIndex={-1}
+            >
+                <h2>{selected}</h2>
+                <SVG>{arrowDown}</SVG>
+            </DropDownHeader>
 
-      <DropDownList style={{ height: isPaymentOpen ? "192px" : 0 }}>
-        {options.map((option, index) => (
-          <ListItem key={index + "-li"} onClick={onOptionClicked(option)}>
-            <ItemButton key={index} type="button" >
-              {options.find((term) => term.includes(String(option)))}
-            </ItemButton>
-          </ListItem>
-        ))}
-      </DropDownList>
-    </Main>
-  );
+            <DropDownList style={{ height: isPaymentOpen ? "192px" : 0 }}>
+                {options.map((option, index) => (
+                    <ListItem key={`${index  }-li`} onClick={onOptionClicked(option)}>
+                        <ItemButton key={index} type="button" >
+                            {options.find((term) => term.includes(String(option)))}
+                        </ItemButton>
+                    </ListItem>
+                ))}
+            </DropDownList>
+        </Main>
+    );
 }
 
 export default FormDropDown;
 
 FormDropDown.propTypes = {
-  isPaymentOpen: PropTypes.bool.isRequired,
-  handlePaymentClick: PropTypes.func.isRequired,
-  selectedPaymentOption: PropTypes.number,
-  handleChangeSelectedOption: PropTypes.func,
+    isPaymentOpen: PropTypes.bool.isRequired,
+    handlePaymentClick: PropTypes.func.isRequired,
+    selectedPaymentOption: PropTypes.number,
+    handleChangeSelectedOption: PropTypes.func,
 };
