+++ src/components/form-components/FormEntry.jsx
@@ -1,7 +1,5 @@
 import styled from "styled-components";
-import PropTypes from "prop-types";
-import { useEffect, useState } from "react";
-import { useWindowWidth } from "../../hooks/useWindowWidth";
+import { useState } from "react";
 
 export const FormEntryContainer = styled.div`
   display: flex;
@@ -28,43 +26,36 @@ const OppositeWidthContainer = styled(FormEntryContainer)`
 
 // todo double check the mobile layout here
 function FormEntry({ className, isLongOnMobile = false, children }) {
-  const [isDirty, setIsDirty] = useState(false);
-  const windowWidth = useWindowWidth();
-  const handleChange = () => {
-    setIsDirty(true);
-  };
+    const [isDirty, setIsDirty] = useState(false);
+    const handleChange = () => {
+        setIsDirty(true);
+    };
 
-  // mobile view gets a longer component, hence the awkward naming scheme here
-  if (isLongOnMobile) {
-    return (
-      <OppositeWidthContainer
-        onChange={handleChange}
-        isDirty={isDirty}
-        className={className}
-      >
-        {children}
-      </OppositeWidthContainer>
-    );
-  }
+    // mobile view gets a longer component, hence the awkward naming scheme here
+    if (isLongOnMobile) {
+        return (
+            <OppositeWidthContainer
+                onChange={handleChange}
+                isDirty={isDirty}
+                className={className}
+            >
+                {children}
+            </OppositeWidthContainer>
+        );
+    }
 
-  else {
+
     return (
-      <FormEntryContainer
-        onChange={handleChange}
-        isDirty={isDirty}
-        className={className}
-      >
-        {children}
-      </FormEntryContainer>
+        <FormEntryContainer
+            onChange={handleChange}
+            isDirty={isDirty}
+            className={className}
+        >
+            {children}
+        </FormEntryContainer>
     );
-  }
+
 }
 
 export default FormEntry;
 
-FormEntry.propTypes = {
-  children: PropTypes.node,
-  className: PropTypes.string,
-  isLongOnMobile: PropTypes.bool,
-  isAlwaysLong: PropTypes.bool,
-};
