+++ src/components/form-components/LongFormEntry.jsx
@@ -1,7 +1,5 @@
 import styled from "styled-components";
 import PropTypes from "prop-types";
-import { useEffect, useState } from "react";
-import { useWindowWidth } from "../../hooks/useWindowWidth";
 
 export const FormEntryContainer = styled.div`
   display: flex;
@@ -13,27 +11,27 @@ export const FormEntryContainer = styled.div`
 `;
 
 function LongFormEntry({ className, children }) {
-  // const [isDirty, setIsDirty] = useState(false);
-  const handleChange = () => {
+    // const [isDirty, setIsDirty] = useState(false);
+    const handleChange = () => {
     // setIsDirty(true);
-  };
+    };
 
-  return (
-    <FormEntryContainer
-      onChange={handleChange}
-      // isDirty={isDirty}
-      className={className}
-    >
-      {children}
-    </FormEntryContainer>
-  );
+    return (
+        <FormEntryContainer
+            onChange={handleChange}
+            // isDirty={isDirty}
+            className={className}
+        >
+            {children}
+        </FormEntryContainer>
+    );
 }
 
 export default LongFormEntry;
 
-LongFormEntry.propTypes = {
-  children: PropTypes.node,
-  className: PropTypes.string,
-  isLongOnMobile: PropTypes.bool,
-  isAlwaysLong: PropTypes.bool,
-};
+// LongFormEntry.propTypes = {
+//     children: PropTypes.node,
+//     className: PropTypes.string,
+//     isLongOnMobile: PropTypes.bool,
+//     isAlwaysLong: PropTypes.bool,
+// };
