+++ src/hooks/useWindowWidth.tsx
@@ -1,13 +1,13 @@
 import { useState, useEffect } from "react";
 
-export function useWindowWidth() {
-  const [width, setWidth] = useState(window.innerWidth);
+export default function useWindowWidth() {
+    const [width, setWidth] = useState(window.innerWidth);
 
-  useEffect(() => {
-    const handleResize = () => setWidth(window.innerWidth);
-    window.addEventListener("resize", handleResize);
-    return () => window.addEventListener("resize", handleResize);
-  });
+    useEffect(() => {
+        const handleResize = () => setWidth(window.innerWidth);
+        window.addEventListener("resize", handleResize);
+        return () => window.addEventListener("resize", handleResize);
+    });
 
-  return width;
+    return width;
 }
