+++ src/utils/firebase.js
@@ -10,13 +10,13 @@ import { getStorage } from "firebase/storage";
 // Your web app's Firebase configuration
 // For Firebase JS SDK v7.20.0 and later, measurementId is optional
 const firebaseConfig = {
-  apiKey: "AIzaSyBzHXtpWX_xSy_GwoG5nuwDjK8pWUXIZsw",
-  authDomain: "invoice-app-project-51ef3.firebaseapp.com",
-  projectId: "invoice-app-project-51ef3",
-  storageBucket: "invoice-app-project-51ef3.appspot.com",
-  messagingSenderId: "164597558125",
-  appId: "1:164597558125:web:01207b83fd69af896bc864",
-  measurementId: "G-0P0YYQ11QY",
+    apiKey: "AIzaSyBzHXtpWX_xSy_GwoG5nuwDjK8pWUXIZsw",
+    authDomain: "invoice-app-project-51ef3.firebaseapp.com",
+    projectId: "invoice-app-project-51ef3",
+    storageBucket: "invoice-app-project-51ef3.appspot.com",
+    messagingSenderId: "164597558125",
+    appId: "1:164597558125:web:01207b83fd69af896bc864",
+    measurementId: "G-0P0YYQ11QY",
 };
 
 // Initialize Firebase
